refactor: removed _types folder and moved its contents to the top level
