+++ startrak-cl/_app/consoleapp.py
@@ -1,9 +1,9 @@
 import sys
-from _types.process.protocols import Executor, Parser, STException
-from _types.process import parsers as parser
-from _types.process import executors as execs
-from _types.alias import LanguageMode
-from _types.streams import ConsoleInput, ConsoleOutput
+from _process.protocols import Executor, Parser, STException
+from _process import parsers as parser
+from _process import executors as execs
+from alias import LanguageMode
+from streams import ConsoleInput, ConsoleOutput
 import startrak
 _PREFIXES = {'st': '[ST]: ', 'py' : '[PY]: ', 'sh' : '[SH]: ' }
